[ipv6] Implement source address selection
Previously, ipxe would always choose the first usable address for IPv6 routing. This
is problematic when an interface has a globally routable and a ULA address. This patch
partially implements the source address selection algorithm from RFC 6724, section 5.

Since ipxe's IPv6 implementation is incomplete, only rules 1, 2, 6 and 8 of this
algorithm are applicable in this case.